issue #147 refactor: move components to each folder
