+++ src/pages/LandingPage/LandingPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import LandingHeader from "@/components/Landing/LandingHeader";
-import LandingMain from "@/components/Landing/LandingMain";
+import LandingHeader from "@/components/Landing/LandingHeader/LandingHeader";
+import LandingMain from "@/components/Landing/LandingMain/LandingMain";
 
 import { LandingPageContainer } from "./LandingPage.styles";
